refactor models structure
